+++ src/server/rpc/factory.ts
@@ -1,0 +1,5 @@
+import { IRpcService, IRpcServiceOptions, RpcService } from './service';
+
+export function createRpcService(options: IRpcServiceOptions = {}): IRpcService {
+    return new RpcService(options);
+}
